fix bignum/bigint conversion for kudos send

+++ app/assets/v2/js/pages/kudos_send.js
@@ -272,7 +272,7 @@ $(document).ready(function() {
 
     // get kudosPrice from the HTML
     kudosPriceInEth = parseFloat($('#kudosPrice').attr('data-ethprice'));
-    kudosPriceInWei = new web3.utils.BN((kudosPriceInEth * 1.0 * Math.pow(10, 18)));
+    kudosPriceInWei = new web3.utils.BN((BigInt(kudosPriceInEth * 1.0 * Math.pow(10, 18))));
 
     var formData = {
       email: email,
@@ -548,7 +548,7 @@ function sendKudos(email, github_url, from_name, username, amountInEth, comments
           console.log('destinationAccount:' + destinationAccount);
 
           var kudosPriceInEth = parseFloat($('#kudosPrice').attr('data-ethprice')) || $('.kudos-search').select2('data')[0].price_finney;
-          var kudosPriceInWei = new web3.utils.BN((kudosPriceInEth * 1.0 * Math.pow(10, 18)));
+          var kudosPriceInWei = new web3.utils.BN((BigInt(kudosPriceInEth * 1.0 * Math.pow(10, 18))));
 
           if (is_direct_to_recipient) {
             // Step 9
